requirements.txt
tweepy
openai
google-generativeai
google-genai
fastapi
uvicorn
python-multipart
python-dotenv
pytest
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
notion-client
APScheduler
